+++ client/websocket/src/client-websocket.ts
@@ -1,3 +1,3 @@
+export { RemoteError } from "@serve-tools/client-realtime";
 export * from "./lib/connect.js";
-export * from "./lib/RemoteError.js";
 export type * from "./lib/types.js";