+++ server/websocket/src/crossws.ts
@@ -1,7 +1,7 @@
 import type { Protocol, ProtocolDefinition } from "@serve-tools/realtime-protocol";
 import { offersWebSocketSubprotocol, subprotocol } from "@serve-tools/realtime-protocol";
+import { createConnection } from "@serve-tools/server-realtime";
 import type { Hooks, Message, Peer } from "crossws";
-import { createConnection } from "./lib/connection.js";
 import type * as T from "./lib/types.js";
 import type { Awaitable, Connection, ConnectionOptions, Handlers } from "./lib/types.js";